refactors searchTrends function into a new page
